Add queryAll function documentation
A documentation block for the function queryAll has been added into the file requestGeneric.php. This clarification provides information on its operation (retrieving all records from a specified database table), parameters required (a database connection object and the table name), and the return type (an array of all the records).